drbd: merge consecutive discard requests during resync
When DRBD resyncs large areas of a thinly provisioned backing device,
it requests data in blocks of rs-discard-granularity from the resync
source node. Up to now DRBD was issuing discards in these blocks on
the resync target node.
Neither ZFS nor physical devices like that. It seems that discard
requests never queue/pipeline. It causes bad resync performance.

Mitigate that by merging the small discard requests on the resync
target node before issuing them to towards the backing device.
With this change the resync-target node merges discard requests when
they are adjacent. It does that by merging the peer_request objects on
the peer_device->resync_requests list.

Note: The "recv_order" link is actually "request_order" in this case.