drbd: send dagtag after establishing connection
This causes a dagtag to be sent even when no writes occur.

This is necessary to allow online verify to proceed when
DRBD_FF_RESYNC_DAGTAG is in use. We, or some other node, may send a
request to this peer that is dependent on a dagtag for a write that it
did not receive. Prior to this change, if no writes are being issued,
then this peer will wait indefinitely for this dagtag.

In other words, a peer may have the data from a write via a resync, even
when it did not receive the write directly. We need to inform this peer
that it has this data.

Fix this by sending a dagtag whenever we enter a state where we would
send a write to this peer.