drbd: set bits for pending writes before sending bitmap
This solves the same problem as drbd_flush_requests() did, but without
introducing a deadlock.

The deadlock occurred when there were requests suspended due to loss of
quorum. The function drbd_flush_requests() waited for the requests to
reach drbd_req_destroy(). Since they were suspended, this did not occur
until after a state change. The state change waited for the work after
the state change to complete, which did not occur because the resource
worker was still stuck in w_after_state_change(), specifically in
drbd_flush_requests().