drbd: free page_chains of peer_requests early
peer_requests on a resync source node contain the data in a page-chain. After
sending the data to the resny target node the data is no longer needed.
Also, after sending a P_RS_DEALLOCATED the data is no longer needed.

In case the connection has higher delay, or the resync target is slow with
sending back write acks, we can get pressure on the page pool if we hold
on these page-chains too long.

Free them as soon as possible. (While keeping the peer_requests around)

This is a preparation step for the next change.