drbd: reorganize the retry loop in drbd_set_role()
From long time ago drbd_set_role() did a final retry of the promote just
to log the error. Removing this behaviour for cases where we do not
want to retry at all (see next commit).
While doing so, ensure that it retries exactly once if it gets a
SS_PRIMARY_NOP (wants to be primary but peer disk states are not Outdated)
return code.