drbd: outdate a joining node's disk when connecting to a diskless primary
Consider a cluster with a primary diskless node, that has connections
with two diskfull nodes. One of the diskful nodes leaves the cluster
for some time. When it comes back it connects to the diskless node first.
This connection should be accepted by the primary diskless, and the
joining node's disk needs to be downgraded below UpToDate at this
moment.

Downgrading it to Consistent is sufficient as then the diskless primary
will never consider this node for IO.

Obviously such a node gets out of the Consistent disk state as soon
as it finishes the resync form one of the UpToDate diskfull nodes.