drbd: release activity log reference when write for 8.4 peer complete
The commit
15f2580aec52 drbd: respect peer writes held due to conflicts when barrier received
removed a special case for compatibility with 8.4 peers. As a result,
activity log references are now activated for writes from such peers.
Unfortunately, they were never released, because these peers do not send
peer acks.

Fix this by releasing the extents when the write is complete. This is
the same behavior as DRBD 8.4 itself when it is using the activity log
on the Secondary node.