drbd: second fixup to the discard merging code
Setting the LIMITED_by bit is only valid if the limiting neighbour is
already received. In other words: The not set EE_TRIM could mean its
not a trim or that we do not know yet, since the rely is not here yet.

After successful merge with a neighbor we do not need to look further
if we need to set the LIMITED_by bit since that request will be
received as well in the future. (If it was received before it has its
LIMIBTED_by bits already set)

The final LIMITED_BEHIND gets set by the make_resync_request() function.
(Therefore it needs to be conserved in case of a merge)

The LIMITED_FRONT here in the _try_merge_rs_discard() function. That
seems less effort than setting it in make_resync_request().  (This is
the reason it is not necessary to conserve it in case of a merge)

Co-developed-by: Joel Colledge <joel.colledge@linbit.com>