drbd: add clean up on namespace removal
When a network namespace is removed, we need to disconnect all connections
with paths in that specific namespace. Otherwise we might leave some sockets
in a non-existing namespace behind. In addition, we would have dangling
pointers in drbd_paths structs.

While we could try re-homing the paths in the init_net, this is almost always
going to fail, as the namespaces will almost certainly not have a compatible
address assigned.

So in short, when a namespace is removed:
* First force disconnect all the connections with paths in that namespace.
* Remove all the path configurations for those connections.

Care must be taken that we don't "sleep" while holding any locks. To this end,
all connections we need to reconfigure are added to a separate list. With this
list, we are free to wait for the connection state to reach StandAlone, after
which we can remove all paths from the connections configuration.