drbd: protect connections from other net namespace
Ensure that once a network namespace is assigned to a connection (by
adding at least one path), only netlink requests from that namespace
can modify the connection.

For ease of administration, we want so have exactly one associated
network namespace per connection. However, we also want to be able
to "move" connections from one namespace to the other.

To that end, connections get the namespace assignment from drbd_path:
* If no path is configured, the connection is free to be assigned to
  any namespace, namely by adding a path in that namespace.
* If at least one path is configured, the namespace of that path is
  the namespace of the connection.

These rules were not enforced up to now. With a few guards on

* add-path
* del-path
* connect
* disconnect

we can ensure that only requests from the right namespace are able
to change the resource state.