drbd: fix corner case of last issued resync peer_req
The resync requests are issued at a specific rate. It might happen,
when we get a reply for the request that is the last on the list.

When _try_merge_rs_discard() is called for the last on the list, it
can not merge with the entry behind, nor can it set the
EE_RS_TRIM_LIMITED_BEHIND on that last entry.

So, when we get the reply for the next element, we check if the
previous one might lack the EE_RS_TRIM_LIMITED_BEHIND here

For this change, the _try_merge_rs_discard() function needs the
ability to raise the attention of multiple requests. For that, I
rename the remove_list to work_list.