drbd: hold read-lock on uuid_sem until state change commited
Looks like while working on this uuid_sem code (and ironing out deadlocks)
the thinking was that we need to hold this lock for sending and comparing
the uuid set.
Well, it turns out we should hold it until the resulting state change is
committed. (Seems we should release it before resyncs start, looks like there
are interactions with the bitmap lock)

Why should we hold it until the state change got committed?
When the need arises to create a new-current-uuid (a write is issued) just
before the connect, the new-current-uuid gets delayed after the handshake is
completed (with that the write blocks for that time).

When the new-current-uuid happens just before the state changes commit, it
still has no connection to that peer, and the new-current-uuid stays on the
local node only. -> Causing issues later.