compat: netns with old sock_create_kern
sock_create_kern used to not have a parameter for the net namespace. We
have to set the namespace after we created the socket, using a special
function that does not increment the refcount for the namespace: drbd sockets
should never keep the namespace alive, instead the sockets are cleaned up when
the namespace is removed.

We also can't use the normal function to release the socket, instead we have to
use a special version that does not decrease the refcount for the namespace.