compat: fix checking for kvfree()
This fixes compat for 4.18.0-372.19.1.el8_6, which moved kvfree() from
linux/mm.h to linux/slab.h.

This commit adds an include for linux/slab.h to the relevant test, so that
the check correctly finds the function.

At usage site we already include linux/slab.h, so nothing to do in the actual
code.