drbd: do not submit data after receiving less than expected
This fixes an issue where DRBD could write arbitrary data to the backing
disk if the connection is lost at the wrong moment. In particular, while
transmitting a P_DATA packet, and part of the way through a 4K (receiver
page size) block. In this situation DRBD would submit a write of the
originally intended size, even though it had only received part of the
data. The remaining data simply consisted of whatever was in the buffer
page prior to receiving the write. This is likely to be the contents of
a previous write, but could also be whatever was in that page before
DRBD allocated it.

This issue was introduced in commit
da2dae781cea drbd: Move the receive buffer to the transport
This change does not take into consideration that dtt_recv_short() might
receive fewer bytes than requested. The issue affects all releases of
DRBD 9.

It is possible that kernel_recvmsg() and thus dtt_recv_short() might
return less than requested size even though the connection is still up.
However, this does not happen in practice for the single pages that we
request. So there is no need to try to receive the data again. Just
check once at the end for minimal performance impact in this hot-path
function.