drbd: allow deletion of connection with unset ns
When removing a connection, we check that the configured namespace of the
path matches the namespace of the requesting socket. We did not account
for the situation that the connection does no longer have a path configured,
in which case the check for matching namespaces would always fail.

Such a situation is easy to create:
1. Create the resource in a namespace
2. Delete the namespaces
3. Remove a connection from the config and run drbdadm adjust

To fix, we allow disconnects and, as a consequence, deletion of the connection
if either the namespaces match, or the connection does not have a namespace
assigned. Same logic is already applied when adding a path.