docker,focal: allow building HWE kernels from source
In 024e6350f52b15ecac930bbe239f4fcaf70fdc54 we added elfutils to
jammy to fix builds of the 5.15 kernel included in ubuntu jammy. That
kernel is available in ubuntu focal as a HWE kernel, so we actually
already support building it, we just need to add necessary tools in
the focal image, too.