drbd: update req_not_net_done safely
It was possible for the caching pointer "req_not_net_done" to skip a
request. If the connection was lost at that point, the remaining
requests would be marked as RQ_NET_DONE, but this one would be stuck
forever. This occurred when the caching pointer was advanced between:
drbd_process_write_request(req);
and
list_add_tail_rcu(&req->tl_requests, &resource->transfer_log);

This bug was introduced by
d6a86c27605e drbd: skip done requests in tl_walk for connection
In particular, the change to the condition defining req_not_net_done.

In more detail:
* Request r0 has been sent, but not yet barrier-acked,
  req_not_net_done == r0.
* r1 is submitted.
* drbd_send_and_submit(r1) reaches drbd_process_write_request(); this
  calls down to set_cache_ptr_if_null(req_not_net_done, r1), which has
  no effect because req_not_net_done is not NULL.
* Another CPU handles the barrier ack for r0;
  advance_cache_ptr(req_not_net_done, r0) finds no further requests in
  the transfer log and sets req_not_net_done to NULL.
* drbd_send_and_submit(r1) adds r1 to the transfer log.
* Now there is a non-NET_DONE request r1 in the transfer log, while
  req_not_net_done is NULL.

This then leads to problems, for instance:
* r2 is submitted.
* drbd_send_and_submit(r2) reaches drbd_process_write_request();
  this calls down to set_cache_ptr_if_null(req_not_net_done, r2),
  which sets req_not_net_done = r2.
* Now there is a non-NET_DONE request r1 in the transfer log that is
  behind req_not_net_done.
* The connection is lost, tl_walk() marks r2 with NET_DONE, but r1
  remains forever without NET_DONE.

Avoid this by adding the request to the transfer log and then setting
RQ_NET_QUEUED. As a result, set_cache_ptr_if_null() will take care of
setting the caching pointer if required.

In order to implement this, we need to set some net flag for peers to
which out-of-sync should be sent. Use RQ_NET_PENDING for this purpose.
This flag remains set only until the out-of-sync packet has been sent.