drbd: pair READ_ONCE() with WRITE_ONCE()
updating req->net_rq_state is protected by the rq_lock on the object.
In advance_cache_ptr() it it reads req->net_rq_state using thre
READ_ONCE() macro (without holding the lock).

Make the updating side with an WRITE_ONCE(). It does not change the
produced code on x86_64, but I think it improves clarity when the
READ_ONCE()/WRITE_ONCE() are paired.