drbd: only attempt to determine size when connecting if we have a disk
The function drbd_determine_dev_size() should only be called when we
have a backing disk. Expand the scope of the get_ldev() reference count
check in receive_state() to ensure that this is the case.

This fixes an oops that could occur when two nodes are connected, one is
Primary, and an additional DRBD volume/device is added to the resource
on both nodes. Example logs from this oops:

[   22.543053] drbd res/1 drbd2 d-150.test: self 0000000000000000:0000000000000000:0000000000000000:0000000000000000 bits:0 flags:0
[   22.546613] drbd res/1 drbd2 d-150.test: peer's exposed UUID: 0000000000000000
[   22.548526] BUG: unable to handle kernel NULL pointer dereference at 0000000000000010
...
[   22.557083] RIP: 0010:drbd_determine_dev_size+0x5a/0x520 [drbd]
...
[   22.580543] Call Trace:
[   22.582456]  receive_state+0x579/0xd90 [drbd]

Reported by Smithx10 on GitHub.