drbd_transport_rdma: Keep reference after tx_timeout
The comment says that my experience was that no RDMA_CM_EVENT_DISCONNECTED
gets delivered after a tx timeout. It might be a different kernel or a
different reason for the tx-timeout, but now I get these events when
transmitting times out.

A way to be prepared for both cases is:
We drop the reference in the dtr_tx_timeout_work_fn() and clear the
DSB_CONNECTED bit. In case the RDMA_CM_EVENT_DISCONNECTED gets delivered
we only drop the the reference if we have the DSB_CONNECTED bit.

This way we avoid dropping the references too early in case the event gets
delivered. (It leads to crashes while freeing the rx-descs)