drbd: do not warn about not sending based on unset timestamp
The timestamp pre_send_jif is, as the name might suggest, only set when
the request is sent. So we cannot use it with requests that have not yet
yet been sent.

This led to spurious warnings of the form:
... We did not send a request for <time-since-boot>ms ...

This warning has been present since the following commit that changed
the meaning of req_not_net_done to include requests which have not yet
been sent:
d6a86c27605e drbd: skip done requests in tl_walk for connection