drbd_transport_rdma: Split cm->state into bits
Several states were only set but never tested for.
Removed those. After that, we have three relevant states.
CONNECT_REQ, CONNECTED and ERROR. With these as bits in the
state word, we can distinguish between an ERROR while connected and
an error on an unconnected QP.

This is also a preparation step for the next patch