drbd: do not swap sync parameters when connecting
When connecting DRBD sends the P_SYNC_PARAM89 packet. This caused it to
swap certain options with the peer.

For example, with nodes A and B, checking the c-max-rate value using
"drbdsetup show":
* c-max-rate on A == 40M, c-max-rate on B == 40M
* Run on A: drbdadm peer-device-options all --c-max-rate=10M
* c-max-rate on A == 10M, c-max-rate on B == 40M
* Run on A: drbdadm disconnect all ; drbdadm connect all
* c-max-rate on A == 40M, c-max-rate on B == 10M

Note that A and B have swapped the configuration for this option. This
behavior is counter-intuitive.

It appears that this situation is partially caused by the omission of a
call to drbd_send_sync_param() in drbd_adm_peer_device_opts(). We could
add this in, but it wouldn't fix the fact that the values are swapped if
they were differently configured before connecting.

Make the behavior simple and easy to understand by not attempting to
agree on the sync parameters with the peer. Each node simply uses its
own configuration. These options affect the sync target, so in effect we
use the configuration on the sync target to control the resync.