drbd: always use wake_up_all() for state_wait
Lets assume that DRBD is establishing a connection. The node with the lower
node ID sends a P_TWOPC_PREPARE to its peer. Let's assume that peer node
is busy with something else at the moment and sends back a P_TWOPC_RETRY.

The initiator of the two-phase-commit receives that in got_twopc_reply()
and does a wake_up_all(&resource->state_wait).

We know that the worker is in change_cluster_wide_state() sleeping in
wait_event_interruptible_timeout(resource->state_wait,
				 cluster_wide_reply_ready(resource),
				 twopc_timeout(resource));

But there might be other tasks waiting on state_wait as well. Just one
example is drbd_open(): With !(mode & FMODE_WRITE) it can end up in
wait_event_interruptible_timeout(resource->state_wait,
				ro_open_cond(device) != -EAGAIN,
				resource->res_opts.auto_promote_timeout * HZ / 10);
and ro_open_cond() waits a bit more when a connection is C_CONNECTING.

So, that means the task waiting in drbd_open() can consume the wake-up, and
the two-pahse-commit hangs until the timeout elapses.

The solution: Use wake_up_all() always when waking up state_wait.