drbd: fix propagation of MDF_NODE_EXISTS
Unfortunately, the commit message of
5292943d6311 (March 2022)
does not explain why sending the bitmap-UUID for all unassigned slots
is a good idea. It is for sure not, since the receiving side sets the
MDF_NODE_EXISTS flag based on that. Before this fix, all nodes
had MDF_NODE_EXISTS set as soon as a day0 UUID was created.
(=a node gets promoted to primary for the first time)

I guess what that commit intended is to send it also for diskless
peers. In other words, for any peer we have a connection for.

This fix influences the quorum calculation, as this tests the
MDF_NODE_EXISTS bit.