drbd: delay promotion until try_become_up_to_date() has run
This fixes a bug which caused an unexpected connection loss after a
Primary is isolated and one of the remaining nodes promotes quickly
afterwards. The characteristic portion of the log looks like:

19:24:49 n25 drbd res: role( Secondary -> Primary )
19:24:49 n25 drbd res: State change failed: Need access to UpToDate data
19:24:49 n25 drbd res/0 drbd100 n21: Failed: pdsk( UpToDate -> Consistent )
19:24:49 n25 drbd res n21: conn( Connected -> Disconnecting ) peer( Secondary -> Unknown )
19:24:49 n25 drbd res/0 drbd100: quorum( yes -> no )
19:24:49 n25 drbd res/0 drbd100 n21: pdsk( UpToDate -> DUnknown ) repl( Established -> Off )
19:24:49 n25 drbd res n21: error receiving P_STATE, e: -5 l: 0!

The steps to trigger this bug are as follows:
* Nodes A, B and C, all with disk and connected.
* Node B is Primary.
* Node B is isolated from network.
* Nodes A and C each notice that B is no longer reachable and change
  their disk state to Consistent until they can determine whether they
  are UpToDate or Outdated.
* Node C is promoted. This succeeds because C has not yet received the
  P_STATE packet from A which would indicate that A is no longer
  UpToDate. That is, C still has a peer-disk state of UpToDate for A.
* The P_STATE packet from A arrives at C. Since C is Primary and its
  disk is still Consistent, this state change is rejected. C drops its
  connection to A.

A natural fix for this would be to ensure that node A sends its P_STATE
packet before acknowledging the two-phase commit for C's promotion.
However this does not help because P_STATE is sent on the data socket,
while P_TWOPC_YES is sent on the control socket. So they can arrive in
either order.

Instead we fix it by delaying promotion until try_become_up_to_date()
has run. That way we at least know the new state of our own disk, which
is sufficient in this case.