drbd: Fix diskless nodes joining a quorate but degraded partition
This is because a freshly created diskless peer has non of
MDF_HAVE_BITMAP | MDF_NODE_EXISTS set. MDF_NODE_EXISTS gets set at the
moment the connection is established for the first time.

Before this fix, the quorate but degraded partition lost quorum
at that moment, because the connection to the newly added diskless
peer is not up yet.