drbd: diskless nodes inherit quorum from quorate D_UP_TO_DATE peers
In a partition where one node has quorum, all nodes should have quorum.
One could argue it makes only sense for the nodes that can reach a majority
of the nodes with their writes.
For now, implementing the idea of inheriting quorum from neighbours only
for diskless nodes.

This does not replace the current neighbour counting algorithm. Because
think of this examle

  D -- L -- D  |  D

A 4 node cluster, with 3 nodes with disk and one diskless node. It is split
into two partitions. The diskful nodes in the left partition do not have
quorum, but the diskless comes to 2 out of 3 and has quorum with that.

PS:
  The last chunk of this patch is a generic bug fix. We may not send the
  state into arbitrary connections. Because it might be a connection that
  just established its transport before finishing the connect
  two-phase-commit (cstate is C_CONNECTING still).
  It comes up now, because on a node that establishes multiple connections
  concurrently, it is now very likely that quorum changes with the first
  established and that can easily interfere withe the second connect process.