drbd: forward migration for wrongly set MDF_NODE_EXISTS flag
Between 9.1.7 and 9.1.12 drbd was setting MDF_NODE_EXISTS for all peers.
With that the flag got useless. It is a meta-data flag that persists.
Clear it for all not configured nodes if we find it in every peer slot.

This is necessary to make the recent commit 8cad30010
'drbd: Fix diskless nodes joining a quorate but degraded partition'
compatible with meta-data that has the MDF_NODE_EXISTS flag set on
all peer slots.

The only downside of this patch is, that it breaks clusters of 32 nodes.
(And when not all of those nodes are configured)