drbd: outdating indirectly reachable nodes upon promote
When committing a promotion (which happens after phase two), we know
that indirectly reachable nodes with disk have lowered their disk
state to D_OUTDATED already. It is valid when the initiator of that
two-phase commits keep that knowledge in the disk state for those
currently unconnected peers.

This is extremely useful for quorum considerations. Every node that
is marked as outdated can allow a certain partition to be quorate.
The example is a cluster that has lost the connection between the
two nodes with disk (A and C; B is diskless).

    B
   / \
  /   \
 A--X--C

When A promotes (and with that it knows that C is Outdated), it can
hold on to quorum even when B goes down. (Before this patch, it lost
quorum at that moment)