drbd: exclude requests that are not yet queued from "seen_dagtag_sector"
This fixes a bug which manifested with log lines of the form:
[   26.498839] drbd res peer: BAD! BarrierAck #5 received with n_writes=4, expected n_writes=5!

The cause of the issue is:
c6c1b4d9faea drbd: update req_not_net_done safely

Which breaks assumptions made in:
623d00e3848a drbd: send early barriers only when sender has seen all requests

In particular, since c6c1b4d9faea we can have requests on the transfer
log with RQ_NET_PENDING set, but RQ_NET_QUEUED not yet set. If we
encounter these requests from the sender, then we do not send them and
we stop processing the transfer log. However, we still updated
"seen_dagtag_sector" for this request, and so we could end up sending a
barrier for the epoch for this request.

Do not update "seen_dagtag_sector" for such requests. The sender will be
woken again once the request has been queued, so there is no risk of
failing to send a necessary barrier as a result.

Note that we cannot use "current_dagtag_sector" instead of
"seen_dagtag_sector" for this purpose because the peer may become
diskless. In this situation "current_dagtag_sector" is no longer
updated, but we still need to send the final barrier. In addition, we
cannot change the meaning of "current_dagtag_sector" to include requests
that we did not send, because we need it to ensure that we send a dagtag
packet when we resume sending writes.