compat: always advertise no write_same support
WRITE_SAME was recently removed, and with it the function to decide
whether we support WRITE_SAME.

This is problematic because we now no longer call
blk_queue_max_write_same_sectors, which would advertise (a lack of)
WRITE_SAME support to the block layer.

So, if the kernel allows us to specify max_write_same_sectors, do it.
But always set it to zero, since DRBD does not support WRITE_SAME
anymore in general.