compat: patch back QUEUE_FLAG_DISCARD if it exists
Since QUEUE_FLAG_DISCARD was removed, a non-zero max_discard_sectors was
simply used to hint the block layer towards the fact that discards are
supported.

Before that, we had to explicitly set QUEUE_FLAG_DISCARD to indicate
that support.

So, whenever we set max_discard_sectors to non-zero, set the flag. If we
set it to zero, clear it.

Note that DRBD now *always* uses max_discard_sectors as the *only*
indicator to tell whether or not discards are supported. So even if
QUEUE_FLAG_DISCARD is present on a kernel, it will simply be ignored.
But that is fine because we always keep max_discard_sectors and
QUEUE_FLAG_DISCARD in sync, so it should not matter which one we look
at.