containers: add LB_SKIP to entry.sh
This makes it easier to use the images as part of k8s initContainers.
The idea is to try all the available images until the container distro
matches the host distro until the usual actions (e.g., compile from
source) are executed. For this to work, we have to exit 0 so that k8s
tries the next container and does not fail.

In a usual case LB_SKIP should not be set. Setting it to 'yes' triggers
the described behavior. Setting it to 'no' will then make the container
exit 1 if host and container do not match.

Co-developed-by: alexzhc <alex.zheng@daocloud.io>