compat: invert queue_discard_zeroes_data patch
COMPAT_QUEUE_LIMITS_HAS_DISCARD_ZEROES_DATA is defined only on kernels
<=3.10. This means that we (appear to) use the "old" API in the code
base, and just patch it out with the compat layer for modern kernels.

This is exactly the opposite way of how we want to be doing things.

Instead -- even if it makes things a lot more complicated in the compat
layer -- remove the old API from the code base and patch it back *in*
using a compat patch.