drbd: use blk_queue_max_discard_sectors helper
We currently only set q->limits.max_discard_sectors, but that is not
enough. Another field, max_hw_discard_sectors, was introduced. The
difference is that max_discard_sectors can be changed from user space
via sysfs, while max_hw_discard_sectors is the "hardware" upper limit.

So use this helper, which sets both.

And while we're at it, use the appropriate helper for
blk_queue_max_write_zeroes_sectors too. This currently does the same
thing we were already doing, but that may change in the future.