drbd: fix incompatible enum assignments in drbd_nl
The use of drbd_state_rv and drbd_ret_code in drbd_nl.c is a mess.
drbd_adm_finish takes an unsigned int, and expects a value of one of
the aforementioned enums.
There are many places where drbd_state_rv and drbd_ret_code are used
interchangably, defeating the purpose of an enum.

So, when we are easily able to, stick with the enums as long as possible
and then just pass them to drbd_adm_finish as an integer at the last
second. This is the same approach as upstream commit 4b28f3b448df
(drbd: address enum mismatch warnings), which has a similar goal.

Where it is impractical to keep the values separate, simply treat the
value as an unsigned int right away to avoid assigning across enum
types.

This fixes gcc -Wextra warnings.