compat: test and patch for bdevname
Note that this is perhaps not exactly right.
What we are testing: "is the bdevname function defined?"
What we *should* be testing: "does the %pg format specifier exist?"

But since #2 is impossible to test with a compile test, just do #1
instead. This is fine too, even if it patches bdevname back for a few
kernels that don't technically need it.