drbd: remove plausible_request_size()
The plausible_request_size() function was introduces in 2015 with introducing
support for WRITE_SAME. By now WRITE_SAME support is removed from the kernel
again.

With the recent capability of merging discards originating from a resync
operation it can happen that the combined merge request has up to 128MiB
in size. Well, if there is some request re-ordering it can happen that
it has slightly more than 128MiB. That happens when two parts gets merged
that are already bigger than the rs_discard_granularity.

When this happens the code from 2015 prevents clearing this area from
the bitmap and the resync "hangs".

Fix this by removing plausible_request_size(), and do the check in
drbd_submit_bio() instead.