drbd: enable rs-discards if backend discard_granularity > rs_discard_granularity
Allow that if rs_discard_granularity is a multiple of the discard_granularity
of the backing disk.

We are merging the rs-discard requests as big as possible (with an upper
limit of 128MiB). Then the call chain is:

drbd_submit_rs_discard() ->
  drbd_conflict_submit_resync_request() ->
    drbd_submit_peer_request() ->
      drbd_issue_peer_discard_or_zero_out() ->
        drbd_issue_discard_or_zero_out()

and drbd_issue_discard_or_zero_out() ensures that the issued discards are
compatible with the discard_granularity of the backing block device.