drbd: release uuid_sem only after matching state change
Since
2046c28b53e6 drbd: hold read-lock on uuid_sem until state change commited
it was possible for uuid_sem to be released too early. One effect of
this was that receive_state() found that the HOLDING_UUID_READ_LOCK flag
was clear, and so failed to set the data socket to be blocking. As a
result, the connection would then later be dropped with the message:
drbd res peer: sock_recvmsg returned -11

This problem occurred in the situation where another state change
occurred between setting HOLDING_UUID_READ_LOCK and clearing it. Note
that this period is contained within a two-phase commit, but that does
not entirely prevent other state changes from occurring.

Fix this by only clearing HOLDING_UUID_READ_LOCK and releasing the read
lock on the semaphore when a connection state change is applied. Any
connection state change is sufficient, as explained in the code comment.

Note that we must do this after applying the state change, as explained
in 2046c28b53e6:

"""
Why should we hold it until the state change got committed?
When the need arises to create a new-current-uuid (a write is issued) just
before the connect, the new-current-uuid gets delayed after the handshake is
completed (with that the write blocks for that time).

When the new-current-uuid happens just before the state changes commit, it
still has no connection to that peer, and the new-current-uuid stays on the
local node only. -> Causing issues later.
"""

However, we cannot delay it until the after state change work because
other operations on the worker thread acquire the semaphore write lock.