drbd: retry connecting if connection lost before two-phase commit starts
Prior to this change, losing the underlying connection at this point
resulted in connect_work() applying the "incompat_states" logic. When
the retry case based on rr_conflict or on_susp_primary_outdated did not
apply, this caused the connection to become StandAlone.

This is not what we want in this situation. Change the return value so
that we retry connecting.

Update the drbd-headers reference because the error message for
SS_NEED_CONNECTION has been generalized to make more sense here.