compat: generalize bdev_discard_granularity__no_present.cocci
There is no value in bdev_discard_granularity(bdev) being only replaced
with its compatible equivalent if it is used in an assignment when
declaring a local variable.

Generalize this to fit when bdev_discard_granularity(bdev) is used as
expression.

Also the bdev part might be an expression in itself.