compat: make blk_cleanup_disk optional
In the blk_alloc_disk__no_present compat patch, we expect
blk_cleanup_disk to exist. However, that may not be the case on newer
kernels.
Since we could hit both cases, and the blk_cleanup_disk is just a
put_disk on new kernels anyway, just ignore it if we don't match the
blk_cleanup_disk there.